removed erroneous files in bin, fixed a few more misspellings of quiz, moved QuizService from service to services package
