--- bin/README.md
@@ -1,18 +1,0 @@
-#TEAM 2 Project
-
-##What we are building
-We are creating a learning tool for students studying chemistry. This tool will allow a sandbox environment for students to learn/validate chemical equations.
-
-###Features
-- Full periodic table containing:
- - Element name
- - Atomic Number
-- Full UI allowing user interactivity of each element
- - Users can add elements together to form new compounds
- - Users can validate chemical compounds
-
-###Tech Stack
-The following were leveraged to build this project:
-- Spring Boot 4
-- Angular
-- MongoDB
