--- bin/src/main/java/META-INF/persistence.xml
@@ -1,4 +1,0 @@
-<persistence>
-	<persistence-unit>
-		<Flashcard>edu.metrostate.ics499.team2.model.Flashcard</Flashcard>
-		<Game>edu.metrostate.ics499.team2.model.Game</Game></persistence-unit></persistence>
