--- bin/src/main/resources/application.properties
@@ -1,15 +1,0 @@
-spring.data.mongodb.authentication-database=admin
-#spring.data.mongodb.username=root
-#spring.data.mongodb.password=root
-spring.data.mongodb.database=chemistry
-spring.data.mongodb.port=27017
-spring.data.mongodb.host=localhost
-#spring.data.mongodb.field-naming-strategy= # Fully qualified name of the FieldNamingStrategy to use.
-#spring.data.mongodb.grid-fs-database= # GridFS database name.
-#spring.data.mongodb.repositories.enabled=true
-#spring.data.mongodb.uri=mongodb://localhost/test
-
-# App Config
-server.port=8102
-spring.application.name=Chemistry
-server.servlet.context-path=/user
