--- src/main/java/edu/metrostate/ics499/team2/repositories/QuizeRepository.java
@@ -1,13 +1,0 @@
-//package edu.metrostate.ics499.team2.repositories;
-//
-//
-//import org.springframework.data.mongodb.repository.MongoRepository;
-//import org.springframework.stereotype.Repository;
-//
-//import edu.metrostate.ics499.team2.model.Quize;
-//
-//@Repository
-//public interface QuizeRepository extends MongoRepository<Quize, String>{
-//
-//}
-
