+++ src/main/java/edu/metrostate/ics499/team2/services/QuizService.java
@@ -1,4 +1,4 @@
-package edu.metrostate.ics499.team2.service;
+package edu.metrostate.ics499.team2.services;
 
 import edu.metrostate.ics499.team2.model.Quiz;
 
@@ -7,7 +7,7 @@ public interface QuizService {
 	public Quiz deletQuizById(String id);
     public Quiz getQuestionById(String question);
 	public Quiz getAnswerById(String answer);
-	public Quiz createQuize(Quiz quiz);
+	public Quiz createQuiz(Quiz quiz);
